// frontend/src/api/generated/ResourceDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ResourceDto = { id: string, kind: string, name: string, description: string, quantity: bigint, unit_label: string, provider_user_id: string | null, provider_name: string | null, active: boolean, created_at: string, };